add chat websocket

+++ backend/src/main/java/com/qm/bookstore/qm_bookstore/dto/user/request/UserUpdateRequest.java
@@ -13,5 +13,5 @@ public class UserUpdateRequest {
     String username;
     String password;
     String email;
-    Integer roleId;
+    Long roleId;
 }
